Extract transiency check to Exception utils and other review remarks
