+++ axon-server-connector/src/main/java/org/axonframework/axonserver/connector/util/ExceptionSerializer.java
@@ -17,7 +17,6 @@
 package org.axonframework.axonserver.connector.util;
 
 import io.axoniq.axonserver.grpc.ErrorMessage;
-import org.axonframework.common.AxonNonTransientException;
 
 import static org.axonframework.common.ObjectUtils.getOrDefault;
 
@@ -54,15 +53,4 @@ public abstract class ExceptionSerializer {
         return builder.build();
     }
 
-    /**
-     * Indicates whether the given {@code failure} is clearly non-transient. Non-transient exceptions indicate
-     * that the handling of the message will fail in the same way if retried
-     *
-     * @param failure the exception that occurred while processing a message
-     * @return {@code true} if the exception is clearly non-transient
-     */
-    public static boolean isExplicitlyNonTransient(Throwable failure) {
-        return failure instanceof AxonNonTransientException
-                || (failure.getCause() != null && isExplicitlyNonTransient(failure.getCause()));
-    }
 }
